requirements.txt
# Deep Learning
torch==2.1.0
torchvision==0.16.0

# Computer Vision
opencv-python==4.8.1.78
mediapipe==0.10.8

# Data Processing
numpy==1.24.3
pandas==2.1.3

# ML Tools
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
pydantic-settings==2.1.0

# Utilities
pyyaml==6.0.1

# Testing
pytest==7.4.3
httpx==0.25.2

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1
